+++ src/group/group.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@nestjs/common';
-import { GroupRepository } from './group.repository';
+
 import { CreateGroupDto } from './dto/create-group.dto';
-import { Group } from './entities/group.entity';
 import { UpdateGroupDto } from './dto/update-group.dto';
-import * as XLSX from 'exceljs';
+import { Group } from './entities/group.entity';
+import { GroupRepository } from './group.repository';
 
 @Injectable()
 export class GroupService {
@@ -73,66 +73,5 @@ export class GroupService {
 
   async deleteAll(): Promise<void> {
     return this.groupRepository.deleteAll();
-  }
-
-  async convertExcelToJSON(groupId: string, excel: Express.Multer.File) {
-    const workbook = new XLSX.Workbook();
-    await workbook.xlsx.load(excel.buffer);
-
-    const worksheet = workbook.getWorksheet(1);
-
-    const columns = [];
-    worksheet.getRow(1).eachCell((cell) => {
-      columns.push(cell.value);
-    });
-
-    const data = [];
-    worksheet.eachRow((row, rowNumber) => {
-      if (rowNumber !== 1) {
-        const rowObject = {};
-        row.eachCell((cell, colNumber) => {
-          const col = columns[colNumber - 1];
-          rowObject[col] = cell.value;
-        });
-        data.push(rowObject);
-      }
-    });
-
-    return {
-      groupId,
-      data,
-    };
-  }
-
-  async convertMemberExcelToJSON(excel: Express.Multer.File) {
-    const workbook = new XLSX.Workbook();
-    await workbook.xlsx.load(excel.buffer);
-
-    const worksheet = workbook.getWorksheet(1);
-
-    const columns = [];
-    worksheet.getRow(1).eachCell((cell) => {
-      columns.push(cell.value);
-    });
-
-    const data = [];
-    worksheet.eachRow((row, rowNumber) => {
-      if (rowNumber !== 1) {
-        const rowObject = {};
-        const memberInfo = {};
-        row.eachCell((cell, colNumber) => {
-          const col = columns[colNumber - 1];
-          if (col == 'name' || col == '이름') {
-            rowObject['name'] = String(cell.value).trim();
-          } else {
-            memberInfo[col] = String(cell.value).trim();
-          }
-        });
-        rowObject['memberInfo'] = memberInfo;
-        data.push(rowObject);
-      }
-    });
-
-    return data;
   }
 }
